+++ R/check_template.R
@@ -28,9 +28,14 @@ get_runner_hash = function(fname) {
 }
 
 check_gitlab_runner = function(fname1, fname2) {
+  msg = yellow(glue("{symbol$circle_filled} Checking gitlab-ci.yml"))
+  message(msg)
   if (get_runner_hash(fname1) !=  get_runner_hash(fname2)) {
-    stop("gitlab-ci.yml has changed.", call. = FALSE)
+    msg = red(glue("{symbol$cross} gitlab-ci.yml has changed."))
+    message(msg)
+    return(FALSE)
   }
+  return(TRUE)
 }
 
 get_r_template_fnames = function(template_repo_loc) {
@@ -63,14 +68,18 @@ check_template = function(type = "r") {
     # Feedback loop if we test template on it's on master
     return(invisible(NULL))
   }
+
+  message(yellow(symbol$circle_filled, "Cloning template repo"))
   tmp_dir = tempdir()
   on.exit(unlink(tmp_dir))
   template_repo_loc = file.path(tmp_dir, "template")
-  system2("git", args = c("clone", "--depth", "1",
+  system2("git", args = c("clone",
+                          "--depth",
+                          "1",
                           "git@gitlab.com:jumpingrivers-notes/template.git", #nolint
                           template_repo_loc))
-  message(template_repo_loc)
 
+  message(yellow(symbol$circle_filled, "Checking template files"))
   proj_name = get_project_name() #nolint
   # Ensure Rproj are given sensible names - not just notes.Rproj
   fnames = c(glue("notes/notes_{proj_name}.Rproj"),
@@ -81,11 +90,11 @@ check_template = function(type = "r") {
 
   if (type == "r") {
     template_fnames = get_r_template_fnames(template_repo_loc)
-    check_gitlab_runner(".gitlab-ci.yml",
+    runner_check = check_gitlab_runner(".gitlab-ci.yml",
                         file.path(template_repo_loc, ".gitlab-ci.yml"))
   } else if (type == "python") {
     template_fnames = get_python_template_fnames(template_repo_loc)
-    check_gitlab_runner(".gitlab-ci.yml",
+    runner_check = check_gitlab_runner(".gitlab-ci.yml",
                         file.path(template_repo_loc, "python-gitlab-ci.yml"))
   }
 
@@ -94,21 +103,24 @@ check_template = function(type = "r") {
   changed = logical(length(fnames))
   for (i in seq_along(fnames)) {
     changed[i] = has_changed(fnames[i], template_fnames[i])
-    if (isTRUE(changed[i])) {
-      message("File ", fnames[i], " does not match")
+    msg = yellow(glue("{symbol$circle_filled} Checking {fnames[i]}"))
+    message(msg)
+
+        if (isTRUE(changed[i])) {
+      msg = glue("{symbol$cross} File {fnames[i]} does not match")
+      message(red(msg))
     }
   }
-  if (any(changed)) {
-    stop("Files differs from template repo.
+  if (any(changed) || isFALSE(runner_check)) {
+    stop(red("Files differs from template repo.
          Either revert your changes or update the template.
          If you think you want to update the template, make a merge
-         request and ask for feedback.", call. = FALSE)
+         request and ask for feedback."), call. = FALSE)
   }
 
   if (length(list.files(pattern = "*\\.Rproj")) > 0L) {
-    stop("Don't add Rproj files to the base directory.",
-         call. = FALSE)
+    msg = glue("{symbol$cross} Don't add Rproj files to the base directory.")
+    stop(red(msg), call. = FALSE)
   }
-
-
+  message(yellow(symbol$tick, "Template files look good"))
 }