+++ NEWS.md
@@ -1,6 +1,7 @@
 # 0.4.15
   * Update create_final to deal with edge cases
-  
+  * Check for broken URLs
+
 # 0.4.14
   * Check files against cannocial template - check_template()